bytequtes/byte/requirements.txt
asgiref==3.5.2
Django==3.2.4
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4.1
pytz==2022.2.1
sqlparse==0.4.2
zipp==3.8.1
